refactor: recipe 재사용 component 패키지 묶기
