+++ Android/feature/recipe/src/main/java/com/sundaegukbap/banchango/feature/recipe/component/Stars.kt
@@ -1,4 +1,4 @@
-package com.sundaegukbap.banchango.feature.recipe.difficulty
+package com.sundaegukbap.banchango.feature.recipe.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Row
